+++ internal/wasm/jit/arch_arm64.s
@@ -3,8 +3,8 @@
 #include "funcdata.h"
 #include "textflag.h"
 
-// jitcall(codeSegment, ce)
-TEXT ·jitcall(SB),NOSPLIT|NOFRAME,$0-16
+// jitcallImpl(codeSegment, ce)
+TEXT ·jitcallImpl(SB),NOSPLIT|NOFRAME,$0-16
         // Load the address of native code.
         MOVD codeSegment+0(FP),R1
         // Load the address of *callEngine.
