+++ internal/wasm/jit/arch_other.go
@@ -1,0 +1,5 @@
+//go:build !amd64 && !arm64
+
+package jit
+
+type archContext struct{}
